Move dictionaries downloader to tools folder
