+++ novelwriter/tools/dictionaries.py
